requirements.txt
asgiref==3.8.1
Django==5.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.17.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
